requirements.txt
sentence-transformers
scikit-learn
pandas
numpy
streamlit
matplotlib
openpyxl
umap-learn
plotly
